Fixed typing issue where subnets to generate would overflow when given 2^32 subnets to generate. Subnets to generate int switched to unsigned long long

+++ subnetter.cpp
@@ -379,7 +379,7 @@ void VLSM(IP IPAddr, SubnetMask netMask1, SubnetMask netMask2) {
     }
     IP localIPCopy = IPAddr & netMask1;
     int networkMagnitudeDifference = netMask1.hostBits - netMask2.hostBits;
-    int subnetsToGenerate = 1<<networkMagnitudeDifference;
+    unsigned long long int subnetsToGenerate = 1ULL<<networkMagnitudeDifference;
     if (numberFlag < subnetsToGenerate && numberFlag != 0) {
         subnetsToGenerate = numberFlag;
     }
